Extend time now works via the new "predicate" key inside TemplateTableView.
TemplateTableView has removed the valuePredicate option and replaced it with the more flexible predicateValue. This is used like so:

		<dict>
			<key>predicate</key>
			<string>extensionTime = 60 * 60</string>
			<key>destinationPath</key>
			<string>accessoryType</string>
			<key>value</key>
			<integer>3</integer>
			<key>altValue</key>
			<integer>0</integer>
		</dict>

If the predicate evaluates to true then 'value' is used. If predicate evaluates to false then altValue is used. (there are alt versions of all value* keys, value, valuePath and valueFormat).